refactor serum program and instruction data
